Adding 1 row and 1 column to coords values.
Changed coords to return 1 extra row and column in order for
subsequent functions that use numpy subsetting to work
correctly. Numpy subsetting is a closed interval for beginning
but a closed interval for ending of slices.

This is probably not a good solution.  The better solution is for
subsetters to add the extra value, but it works for now.

Also standardized indentation, replacing all tabs with spaces,
and setting indentation to 4 spaces instead of 8.