Replaced coords function with new version.
New version doesn't require a hardcoded year in file to read,
as long as it finds at least 1 file, it will just read the last
one it finds. This should be more versatile and should work no
matter which cubes are available.

This version also adds 1 to x_end and y_end, because the way this
values are generally used by the calling functions, they depend
on an open interval at the end of the subset indices.