Maintenance/upgraded on CETB_analysis.
Deleted obsolete MOD_array_year. User should use MOD_array and
then use the column corresponding to the desired year of MOD
dates. Upgraded MOD_array to return df with 'Avg' column that
contains what used to be passed back as the masked array.
Now the individual years and the average are passed back.
Eventually we should just remove the masked array output.
Fixed a problem with the data frame column names.  Changed the
rows index to be the (row,col) indices.