Fixed numerous issues with 1GRD... notebook.
1GRD notebook is now set up to only call MOD_array and use the df
that is passed back from it. This is set up for Joan to run, but
subsetting is only for a few years because this procedure is a
real memory piggy. Now that we have things working with the data
frame from MOD_array, we should delete that masked array we were
using before. This will allow more years to be handled.

This works through the cell that makes the SIR and GRD plots,
side-by-side, with the GRD MOD inserted as a horizontal bar in
the SIR colorbar.