Added EHD analysis, MOD forward-facing indexer, and new directory output location
EHD (end of high DAV) analysis and output code were added to CETB_analysis script and Large Melt Map with geolocated df and geotiffs.

Added forward-facing indexer line to notebook so that MOD is assigned to the first day in a specified analysis window AND on a day that meets the melt threshold criteria. Melt threshold criteria requirement was implemented in the CETB_analysis script (MOD_array function).

Added a new variable to local machine locations called outDir so that plots and pkl files are not written to shared R drive location.